// src/types/player-stats.ts
export interface TomatoPlayerStats {
  battles: number;
  avgTier: number;
  wn8: number;
  wnx: number;
  wins: number;
  winrate: number;
  survived: number;
  survivalRate: number;
  damage: number;
  avgDamage: number;
  damageRatio: number;
  frags: number;
  avgFrags: number;
  kdRatio: number;
}